python_django_diploma/marketplace/requirements.txt
amqp==5.2.0
api==0.0.7
asgiref==3.6.0
autopep8==2.0.4
billiard==4.2.0
celery==5.3.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.1
django-admin-rangefilter==0.12.1
django-ckeditor==6.7.1
django-filter==23.3
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-mptt-admin==2.4.1
django-phonenumber-field==7.0.2
django-silk==5.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
factory-boy==3.3.0
Faker==22.2.0
gprof2dot==2022.7.29
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.3
jsonargparse==4.27.1
kombu==5.3.5
Markdown==3.4.4
MarkupSafe==2.1.5
nose==1.3.7
packaging==23.2
phonenumbers==8.13.8
Pillow==9.4.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.6
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.17.2
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.28.2
rest-framework-simplejwt==0.0.2
schedule==1.2.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.3
tzdata==2024.1
urllib3==1.26.15
vine==5.1.0
wcwidth==0.2.13
gunicorn==21.2.0
redis==4.6.0
django-modeltranslation==0.18.11